adaptação da tela Descobrir para responsividade e criação da visualização do projeto para mobile
