+++ frontend-orange/src/pages/DescobrirDesktop.jsx
@@ -1,30 +1,32 @@
-import { Box, ThemeProvider, Link, Typography,
-         Button, Modal } from '@mui/material/';
+import { Box, ThemeProvider, Link, Typography} from '@mui/material/';
 import {useMediaQuery} from '@mui/material/';
 import {theme} from '../utils/Theme'
 import Header from '../components/main-header/Header';
-
 import FormBuscarTags from '../components/meu-portfolio/FormBuscarTags';
-
 import { useState } from "react";
+import figuraProjeto from '../assets/projects/project1.svg'
 import profilePicture from '../assets/profile-picture/picture.svg';
 import CardProjeto from '../components/meu-portfolio/CardProjeto';
 import ModalProjeto from '../components/meu-portfolio/ModalProjeto';
 
-
-
-export default function MeuPortfolio(){
+export default function DescobrirDesktop(){
   const responsivo1 = useMediaQuery(theme.breakpoints.up('sm'))
+  
+  const [open, setOpen] = useState(false);
+  const handleOpen = () => setOpen(true);  
+  const handleClose = () => setOpen(false);
 
-  const [cardPerfil] = useState({
+  const [cardPerfil, setCardPerfil] = useState({
     pais:'Brasil',
     avatar: profilePicture,
     nome:'John Doe',
     email:'john@doeux.com',
     tag:{ui:'UI', ux:'UX', web:'Web'},
-    data:'12/23'
+    data:'12/23',
+    urlImagem:figuraProjeto,
+    tituloProj:'Ecommerce One Page',
+    descricao: 'Temos o prazer de compartilhar com vocês uma variação do nosso primeiro recurso gratuito. É um modelo de IA. Tentamos redesenhar uma versão mais minimalista do nosso primeiro projeto.'
   })
-
 
   return(
   <ThemeProvider theme={theme}>
@@ -37,7 +39,7 @@ export default function MeuPortfolio(){
       />
       <Typography 
         sx={{
-          margin:responsivo1 ? '112px 268px' : '24px 55px',
+          margin:responsivo1 ? '112px 268px' : '55px 24px',
           textAlign: 'center'
           }}
           variant={responsivo1 ? "h4" : "h5"}
@@ -53,13 +55,13 @@ export default function MeuPortfolio(){
         </Box>
 
         <Box sx={{display:'flex', gap:'26px', flexWrap:'wrap'}}>
-           <Link
-            onClick={()=>{console.log('clicando')}}
-            underline='none'
-            style={{cursor:'pointer'}}
-           > 
+          <Link
+          onClick={handleOpen}
+          underline='none'
+          style={{cursor:'pointer'}}
+          width={responsivo1 ? 'calc(33.33% - 17.33px)' : '100%'}
+          > 
             <CardProjeto 
-              width={responsivo1 ? 'calc(33.33% - 17.33px)' : '100%'}
               color='neutral.dark'
               colorIconMenu='secondary.secondaryLight'
               avatar={cardPerfil.avatar}
@@ -72,13 +74,33 @@ export default function MeuPortfolio(){
               labelChip2={cardPerfil.tag.ux}
               labelChip3={cardPerfil.tag.web}
               data={cardPerfil.data}
-              iconMenu='none'              
+              iconMenu='none'
+              imagem={cardPerfil.urlImagem}             
               >           
             </CardProjeto>
-          </Link>          
+          </Link> 
+                   
         </Box>
       </Box>
-    </Box>     
+    </Box>
+
+    {open ?  
+    
+      <ModalProjeto
+      open={open}
+      onClose={handleClose}
+      avatar={cardPerfil.avatar}
+      nome={cardPerfil.nome}
+      data={cardPerfil.data}
+      labelChip1={cardPerfil.tag.ui}
+      labelChip2={cardPerfil.tag.ux}
+      tituloProj={cardPerfil.tituloProj}
+      imagem={cardPerfil.urlImagem}
+      descricao={cardPerfil.descricao}
+      
+      />
+    : null} 
+   
   </ThemeProvider>
   )
 }
